+++ golang-backend/internal/domain/project_status.go
@@ -4,7 +4,7 @@ type ProjectStatus struct {
 	ID       string
 	Name     string
 	HexColor string
-	Index    uint
+	Index    uint8
 	Flags    Bitmask
 }
